- spearation of initialisation scripts - cleaning
