+++ energy_demand/scripts/s_generate_sigmoid.py
@@ -112,142 +112,6 @@ def calc_sigmoid_parameters(l_value, xdata, ydata): #, fit_crit_max=400, fit_cri
 
     return fit_parameter
 
-def tech_sigmoid_parameters(
-        base_yr,
-        technologies,
-        enduse,
-        crit_switch_service,
-        installed_tech,
-        l_values,
-        service_tech_by_p,
-        service_tech_switched_p,
-        fuel_switches,
-        service_switches
-    ):
-    """Calculate diffusion parameters based on energy service
-    demand in base year and projected future energy service demand
-
-    The future energy servie demand is calculated based on fuel switches.
-    A sigmoid diffusion is fitted.
-
-    Arguments
-    ----------
-    base_yr : dict
-        base year
-    technologies : dict
-        technologies
-    enduse : str
-        Enduse
-    crit_switch_service : bool
-        Criteria whether sigmoid is calculated for service switch or not
-
-    installed_tech : list
-        List with installed technologies in fuel switches
-    l_values : dict
-        L values for maximum possible diffusion of technologies
-    service_tech_by_p : dict
-        Energy service demand for base year (1.sigmoid point)
-    service_tech_switched_p : dict
-        Service demand after fuelswitch
-    fuel_switches : dict
-        Fuel switch information
-
-    Returns
-    -------
-    sigmoid_parameters : dict
-        Sigmoid diffusion parameters to read energy service demand percentage (not fuel!)
-
-    Notes
-    -----
-    Manually the fitting parameters can be defined which are not considered as
-    a good fit: fit_crit_max, fit_crit_min
-    If service definition, the year until switched is the end model year
-    """
-    # As fit does not work with a starting point of 0,
-    # an initial small value needs to be assumed
-    fit_assump_init = 0.001
-
-    sigmoid_parameters = defaultdict(dict)
-
-    # Fitting criteria where the calculated sigmoid slope and midpoint can be provided limits
-    if installed_tech == []:
-        logging.debug("NO TECHNOLOGY...%s %s", enduse, installed_tech)
-    else:
-        for tech in installed_tech:
-            logging.debug("... create sigmoid diffusion parameters %s %s", enduse, tech)
-
-            # If service switch
-            if crit_switch_service:
-
-                # Get year until switched
-                for switch in service_switches:
-                    if switch.technology_install == tech:
-                        yr_until_switched = switch.switch_yr
-
-                market_entry = technologies[tech].market_entry
-            else: #fuel switch
-
-                # Get the most future year of the technology in the enduse which is switched to
-                yr_until_switched = 0
-                for switch in fuel_switches:
-                    if switch.enduse == enduse and switch.technology_install == tech:
-                        if yr_until_switched < switch.switch_yr:
-                            yr_until_switched = switch.switch_yr
-
-                market_entry = technologies[tech].market_entry
-
-            # --------
-            # Test whether technology has the market entry before or after base year,
-            # If afterwards, set very small number in market entry year
-            # --------
-            if market_entry > base_yr:
-                point_x_by = market_entry
-                point_y_by = fit_assump_init
-            else: # If market entry before, set to 2015
-                point_x_by = base_yr
-                point_y_by = service_tech_by_p[tech] # current service share
-
-                #If the base year is the market entry year use a very small number
-                if point_y_by == 0:
-                    point_y_by = fit_assump_init
-
-            # Future energy service demand (second point on sigmoid curve for fitting)
-            point_x_projected = yr_until_switched
-            point_y_projected = service_tech_switched_p[tech]
-
-            # Data of the two points
-            xdata = np.array([point_x_by, point_x_projected])
-            ydata = np.array([point_y_by, point_y_projected])
-
-            # ----------------
-            # Parameter fitting
-            # ----------------
-            fit_parameter = calc_sigmoid_parameters(
-                l_values[enduse][tech],
-                xdata,
-                ydata)
-
-            logging.debug(
-                " ... Fitting: Midpoint: %s steepness: %s", fit_parameter[0], fit_parameter[1])
-
-            # Insert parameters
-            sigmoid_parameters[tech]['midpoint'] = fit_parameter[0] #midpoint (x0)
-            sigmoid_parameters[tech]['steepness'] = fit_parameter[1] #Steepnes (k)
-            sigmoid_parameters[tech]['l_parameter'] = l_values[enduse][tech]
-
-            #plot sigmoid curve
-            '''from energy_demand.plotting import plotting_program
-            plotting_program.plotout_sigmoid_tech_diff(
-                 l_values[enduse][tech],
-                 tech,
-                 enduse,
-                 xdata,
-                 ydata,
-                 fit_parameter,
-                 False)'''
-
-    return dict(sigmoid_parameters)
-
 def get_tech_future_service(service_tech_by_p, service_tech_ey_p, regions=False, regional_specific=False):
     """Get all those technologies with increased service in future
 
@@ -285,17 +149,14 @@ def get_tech_future_service(service_tech_by_p, service_tech_ey_p, regions=False,
 
     if regional_specific:
         for reg in regions:
+            tech_increased_service[reg] = {}
+            tech_decreased_share[reg] ={}
+            tech_constant_share[reg] = {} 
 
             # If no service switch defined
             if service_tech_ey_p[reg] == {}:
-                tech_increased_service[reg] = []
-                tech_decreased_share[reg] = []
-                tech_constant_share[reg] = []
+                pass
             else:
-                tech_increased_service[reg] = {}
-                tech_decreased_share[reg] = {}
-                tech_constant_share[reg] = {}
-
                 # Calculate fuel for each tech
                 for tech in service_tech_by_p.keys():
                     if tech == 'dummy_tech':
@@ -311,12 +172,13 @@ def get_tech_future_service(service_tech_by_p, service_tech_ey_p, regions=False,
                         else: #same
                             tech_constant_share[reg][tech] = service_tech_ey_p[reg][tech]
     else:
+        tech_increased_service = {}
+        tech_decreased_share = {}
+        tech_constant_share = {}
 
         # If no service switch defined
         if service_tech_ey_p == {}:
-            tech_increased_service = []
-            tech_decreased_share = []
-            tech_constant_share = []
+            pass
         else:
 
             # Calculate fuel for each tech
@@ -415,7 +277,7 @@ def tech_l_sigmoid(
     l_values_sig = {}
 
     # Check wheter there are technologies in this enduse which are switched
-    if installed_tech[enduse] == []: #[enduse] == []:
+    if installed_tech[enduse] == []:
         pass
     else:
         logging.debug("Technologes it calculate sigmoid %s %s", enduse, installed_tech[enduse])
@@ -481,7 +343,6 @@ def calc_service_fuel_switched(
     """
     service_tech_switched_p = {}
 
-    #service_tech_switched_p = defaultdict(dict) #BELUGA SCRAP
     for fuel_switch in fuel_switches:
         if fuel_switch.enduse == enduse:
 
@@ -522,7 +383,7 @@ def calc_service_fuel_switched(
     for tech in service_tech_by_p:
         if tech not in service_tech_switched_p.keys():
             all_fractions_not_affected_by_switch[tech] = service_tech_by_p[tech]
-    
+
     # Iterate all technologies of enduse_by
     service_tot_remaining = sum(all_fractions_not_affected_by_switch.values())
 
@@ -531,15 +392,15 @@ def calc_service_fuel_switched(
 
         # Relative share
         rel_service_fraction_p = all_fractions_not_affected_by_switch[tech] / service_tot_remaining
-        
+
         service_tech_switched_p[tech] = rel_service_fraction_p * remaining_service_to_distr_p
 
     return dict(service_tech_switched_p)
 
 def get_tech_installed_single_enduse(enduse, fuel_switches):
-    """Read out all technologies which are
-    specifically switched to for all enduses
-    TODO CHANGED
+    """Read out all technologies which are specifically switched
+    to for an enduse
+
     Parameter
     ---------
     enduse : str
@@ -561,195 +422,13 @@ def get_tech_installed_single_enduse(enduse, fuel_switches):
     for switch in fuel_switches:
         if switch.enduse == enduse:
             installed_tech[enduse].add(switch.technology_install)
+
     # Convert set to lists
     for enduse, set_values in installed_tech.items():
         installed_tech[enduse] = list(set_values)
 
     return installed_tech
 
-'''def get_sig_diffusion(
-        base_yr,
-        technologies,
-        service_switches,
-        fuel_switches,
-        enduses,
-        tech_increased_service,
-        service_tech_ey_p,
-        service_fueltype_by_p,
-        service_tech_by_p,
-        fuel_tech_p_by,
-        regions=False,
-        regional_specific=False
-    ):
-    """Calculates parameters for sigmoid diffusion of
-    technologies which are switched to/installed.
-
-    Arguments
-    ----------
-    data : dict
-        Data
-    service_switches : dict
-        Service switches
-    fuel_switches : dict
-        Fuel switches
-    enduses : list
-        Enduses
-    tech_increased_service : list
-        Technologies with increased service
-    service_tech_ey_p : dict
-        Fraction of service in end year
-    service_fueltype_by_p :
-        Fraction of service per fueltype in base year
-    service_tech_by_p :
-        Fraction of service per technology in base year
-    fuel_tech_p_by :
-        Fraction of fuel per technology in base year
-    regional_specific : bool
-        Whether the calculation is for all regions or not
-
-    Return
-    ------
-    data : dict
-        Data dictionary containing all calculated
-        parameters in assumptions
-
-    Note
-    ----
-    It is assumed that the technology diffusion is the same over
-    all the uk (no regional different diffusion)
-    """
-    if len(service_switches) > 0:
-        crit_switch_service = True
-    else:
-        crit_switch_service = False
-
-    if regional_specific:
-        installed_tech, sig_param_tech = defaultdict(dict), {}
-        l_values_sig = defaultdict(dict)
-        sig_param_tech = defaultdict(dict)
-    else:
-        installed_tech, sig_param_tech = {}, {}
-        l_values_sig = defaultdict(dict)
-
-    for enduse in enduses:
-
-        if crit_switch_service:
-            """Sigmoid calculation in case of 'service switch'
-            """
-
-            # Tech with lager service shares in end year
-            installed_tech[enduse] = tech_increased_service[enduse]
-
-            # End year service shares (scenaric input)
-            service_tech_switched_p = service_tech_ey_p
-
-            if regional_specific:
-                # Maximum shares of each technology
-                for reg in regions:
-                    l_values_sig[reg][enduse] = {}
-                    #same techs for every region
-                    for tech in installed_tech[enduse][reg]:
-                        l_values_sig[reg][enduse][tech] = technologies[tech].tech_max_share
-            else:
-
-                if tech_increased_service[enduse] == []:
-                    pass
-                else:
-                    # Maximum shares of each technology
-                    for tech in installed_tech[enduse]:
-                        l_values_sig[enduse][tech] = technologies[tech].tech_max_share
-        else:
-            """Sigmoid calculation in case of 'fuel switch'
-            """
-            if regional_specific:
-
-                # Tech with lager service shares in end year (installed in fuel switch)
-                techs = get_tech_installed(enduses, fuel_switches)
-                for reg in regions:
-                    for _enduse in techs.keys():
-                        installed_tech[_enduse][reg] = techs[_enduse]
-
-                service_tech_switched_p = {}
-
-                for reg in regions:
-
-                    # Calculate future service demand after fuel switches for each technology
-                    service_tech_switched_p[reg] = calc_service_fuel_switched(
-                        enduses,
-                        fuel_switches,
-                        technologies,
-                        service_fueltype_by_p,
-                        service_tech_by_p,
-                        fuel_tech_p_by,
-                        techs,
-                        'actual_switch')
-
-                    # Calculate L for every technology for sigmod diffusion
-                    l_values_sig[reg] = tech_l_sigmoid(
-                        enduses,
-                        fuel_switches,
-                        technologies,
-                        techs,
-                        service_fueltype_by_p,
-                        service_tech_by_p,
-                        fuel_tech_p_by)
-            else:
-
-                # Tech with lager service shares in end year (installed in fuel switch)
-                installed_tech = get_tech_installed(enduses, fuel_switches)
-
-                # Calculate future service demand after fuel switches for each technology
-                service_tech_switched_p = calc_service_fuel_switched(
-                    enduses,
-                    fuel_switches,
-                    technologies,
-                    service_fueltype_by_p,
-                    service_tech_by_p,
-                    fuel_tech_p_by,
-                    installed_tech,
-                    'actual_switch')
-
-                # Calculate L for every technology for sigmod diffusion
-                l_values_sig = tech_l_sigmoid(
-                    enduses,
-                    fuel_switches,
-                    technologies,
-                    installed_tech,
-                    service_fueltype_by_p,
-                    service_tech_by_p,
-                    fuel_tech_p_by)
-
-        if regional_specific:
-
-            for reg in l_values_sig:
-                # Calclulate sigmoid parameters for every installed technology
-                sig_param_tech[reg][enduse] = tech_sigmoid_parameters(
-                    base_yr,
-                    technologies,
-                    enduse,
-                    crit_switch_service,
-                    installed_tech[enduse][reg],
-                    l_values_sig[reg],
-                    service_tech_by_p[enduse],
-                    service_tech_switched_p[reg][enduse],
-                    fuel_switches,
-                    service_switches)
-        else:
-            # Calclulate sigmoid parameters for every installed technology
-            sig_param_tech[enduse] = tech_sigmoid_parameters(
-                base_yr,
-                technologies,
-                enduse,
-                crit_switch_service,
-                installed_tech[enduse],
-                l_values_sig,
-                service_tech_by_p[enduse],
-                service_tech_switched_p[enduse],
-                fuel_switches,
-                service_switches)
-
-    return dict(installed_tech), dict(sig_param_tech), dict(service_tech_switched_p) #NEW --> End year service fraction
-'''
 def get_sig_diffusion_service(
         technologies,
         tech_increased_service,
@@ -864,14 +543,14 @@ def calc_diff_fuel_switch(
     all the uk (no regional different diffusion)
     """
     if regional_specific:
-        installed_tech = defaultdict(dict)
+        #installed_tech = defaultdict(dict)
         l_values_sig = defaultdict(dict)
         service_tech_switched_p = {}
 
         # Tech with lager service shares in end year (installed in fuel switch)
         techs = get_tech_installed_single_enduse(enduse, fuel_switches)
-        for reg in regions:
-            installed_tech[reg] = techs
+        ##for reg in regions:
+        ###    installed_tech[reg] = techs
 
         for reg in regions:
 
@@ -930,24 +609,53 @@ def calc_sigm_parameters(
         service_tech_by_p,
         service_tech_switched_p,
         service_switches,
-        techs,
+        tech_increased_service,
         regions=False,
         regional_specific=False
     ):
+    """
+    TODO
+
+    Arguments
+    ---------
+    base_yr : int
+        Base year
+    technologies : dict
+        Technologies
+    l_values_sig : dict
+
+    service_tech_by_p : dict
+
+    service_tech_switched_p : dict
+
+    service_switches : list
+        Service switches
+    tech_increased_service : dict
+        Technologies with increased service in end year
+    regions : dict, default=False
+        Regions
+    regional_specific : bool, default=False
+        Criteria if regional specific calcuations
+
+    Returns
+    ------- 
+    sig_param_tech : dict
+        Sigmoid parameters for increasing technologies
+    """
     if regional_specific:
-        installed_tech = defaultdict(dict)
+        installed_tech = {} #defaultdict(dict)
         sig_param_tech = {}
 
         for reg in regions:
-            if techs[reg] == []:
-                installed_tech[reg] = []
-            else:
-                installed_tech[reg] = techs[reg].keys()
+            #if tech_increased_service[reg] == {}: #[]:
+            #    installed_tech[reg] = []
+            #else:
+            installed_tech[reg] = tech_increased_service[reg].keys()
 
         for reg in l_values_sig:
 
             # Calclulate sigmoid parameters for every installed technology
-            sig_param_tech[reg] = tech_sigmoid_parameteNEW(
+            sig_param_tech[reg] = tech_sigmoid_parameters(
                 base_yr,
                 technologies,
                 installed_tech[reg],
@@ -958,13 +666,13 @@ def calc_sigm_parameters(
     else:
         sig_param_tech = {}
 
-        if techs == []:
-            installed_tech = []
-        else:
-            installed_tech = list(techs.keys())
+        #if tech_increased_service == {}: #[]:
+        #    installed_tech = []
+        #else:
+        installed_tech = list(tech_increased_service.keys())
 
         # Calclulate sigmoid parameters for every installed technology
-        sig_param_tech = tech_sigmoid_parameteNEW(
+        sig_param_tech = tech_sigmoid_parameters(
             base_yr,
             technologies,
             installed_tech,
@@ -975,7 +683,7 @@ def calc_sigm_parameters(
 
     return sig_param_tech
 
-def tech_sigmoid_parameteNEW(
+def tech_sigmoid_parameters(
         base_yr,
         technologies,
         installed_tech,
